requirements.txt
numpy>=1.21.0
google-genai>=1.15.0
colorama>=0.4.6
pytest>=7.0.0
matplotlib>=3.5.0
python-dotenv>=0.19.0
toml>=0.10.2